+++ BeyondStorage/Source/Configuration/Gears/GearsConversions.cs
@@ -1,11 +1,43 @@
 ﻿using System;
+using BeyondStorage.Configuration;
 using BeyondStorage.Infrastructure;
 
 namespace BeyondStorage.Source.Configuration.Gears;
 
 internal static class GearsConversions
 {
-    internal static bool ToBool(string value)
+    public static bool IsEqualValue(string value, bool b)
+    {
+        var a = ToBool(value, !b);
+        return a == b;
+    }
+
+    public static bool IsEqualValue(string value, float b)
+    {
+        var a = ToFloat(value, -b);
+        return a == b;
+    }
+
+    internal static string FromBool(bool value)
+    {
+        return value ? "On" : "Off";
+    }
+
+    internal static string FromFloat(float value)
+    {
+        if (value < ModConfig.RANGE_UNLIMITED)
+        {
+            value = ModConfig.RANGE_UNLIMITED;
+        }
+        else if (value > ModConfig.RANGE_MAX_USER_LIMIT)
+        {
+            value = ModConfig.RANGE_MAX_USER_LIMIT;
+        }
+
+        return value.ToString("F1");
+    }
+
+    internal static bool ToBool(string value, bool defaultValue)
     {
         if (string.Equals(value, "Off", StringComparison.InvariantCultureIgnoreCase))
         {
@@ -18,7 +50,16 @@ internal static class GearsConversions
         else
         {
             ModLogger.DebugLog($"Cannot convert `{value}` to a bool setting value");
-            return true;
+            return defaultValue;
         }
+    }
+    internal static float ToFloat(string value, float defaultValue)
+    {
+        if (!float.TryParse(value, out var convertedValue))
+        {
+            convertedValue = defaultValue;
+        }
+
+        return convertedValue;
     }
 }