+++ BeyondStorage/Source/Configuration/Gears/GearsMod_Init.cs
@@ -5,27 +5,35 @@ using GearsAPI.Settings.World;
 
 namespace BeyondStorage.Source.Configuration.Gears;
 
-public class GearsMod_Init : IGearsModApi
+public class GearsModAPI : IGearsModApi
 {
     private static IGearsMod s_gearsMod;
-    private static IModGlobalSettings s_GearsGlobalSettings;
 
+    private static IModGlobalSettings GearsGlobalSettings
+    {
+        get; set;
+    }
 
     void IGearsModApi.InitMod(IGearsMod modInstance)
     {
         s_gearsMod = modInstance;
     }
 
+    public static void SaveGlobalSettings()
+    {
+        GearsGlobalSettings?.SaveSettings();
+    }
+
     void IGearsModApi.OnGlobalSettingsLoaded(IModGlobalSettings modSettings)
     {
-        s_GearsGlobalSettings = modSettings;
-        if (s_GearsGlobalSettings == null)
+        GearsGlobalSettings = modSettings;
+        if (GearsGlobalSettings == null)
         {
             ModLogger.DebugLog($"Global settings loaded, but modSettings is null, and gears mod is {s_gearsMod}");
             return;
         }
 
-        var generalTab = s_GearsGlobalSettings.GetTab("General");
+        var generalTab = GearsGlobalSettings.GetTab("General");
         if (generalTab == null)
         {
             ModLogger.DebugLog($"Global settings loaded, but generalTab is null");
@@ -33,21 +41,13 @@ public class GearsMod_Init : IGearsModApi
         }
 
         var generalCategory = generalTab.GetCategory("General");
-        if (generalTab == null)
+        if (generalCategory == null)
         {
             ModLogger.DebugLog($"Global settings loaded, but generalCategory is null");
             return;
         }
 
-        var isDebugSetting = (generalCategory.GetSetting("IsDebug") as IGlobalValueSetting);
-        if (isDebugSetting == null)
-        {
-            ModLogger.DebugLog($"Global settings loaded, but isDebugSetting is null");
-            return;
-        }
-
-        isDebugSetting.OnSettingChanged += GearsGeneralSettings.SetDebugMode;
-        GearsGeneralSettings.SetDebugMode(isDebugSetting, isDebugSetting.CurrentValue);
+        GearsGeneralSettings.Configure(generalCategory);
     }
 
     void IGearsModApi.OnWorldSettingsLoaded(IModWorldSettings worldSettings)